+++ src/Featurizers/Components/WindowFeaturizerBase.h
